backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
instaloader==4.13.1
python-dotenv==1.0.1
httpx==0.27.0
pydantic==2.10.6